Move all content to src
