+++ .env.example
@@ -1,1 +1,1 @@
-FIXER_API_KEY=________________
+API_LAYER_KEY=________________
