+++ src/helpers/functional.ts
