+++ src/pages/index.tsx
@@ -1,8 +1,8 @@
 import type { NextPage } from 'next'
 import Link from 'next/link'
 
-import Layout from '../components/Layout'
-import styles from '../styles/Home.module.scss'
+import Layout from 'components/Layout'
+import styles from 'styles/Home.module.scss'
 
 const Home: NextPage = () => {
   return (
